requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.18.0
python-dotenv>=0.19.0
requests>=2.26.0
rdkit>=2022.9.1
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-groq>=0.0.1
langgraph>=0.0.1
pydantic>=2.0.0